refactor(ui): extract custom trace helper
