+++ test/ui/fixtures/trace-custom/basic.test.ts
@@ -1,10 +1,6 @@
-import type { Page } from 'playwright'
-import { createRequire } from 'node:module'
 import { chromium } from 'playwright'
-import { onTestFinished, recordArtifact, test, vi } from 'vitest'
-
-const require = createRequire(import.meta.url)
-const rrwebSnapshotPath = require.resolve('rrweb-snapshot')
+import { onTestFinished, test } from 'vitest'
+import { createTraceRecorder } from './trace'
 
 test('custom trace', async ({ task }) => {
   const browser = await chromium.launch()
@@ -12,49 +8,12 @@ test('custom trace', async ({ task }) => {
 
   const page = await browser.newPage()
   await page.setContent('<main><button>Before action</button></main>')
-  await page.addScriptTag({ path: rrwebSnapshotPath })
+  const trace = await createTraceRecorder(page, task)
 
-  await recordTrace(page, task, 'before action', 0)
+  await trace.snapshot('before action')
 
   await page.locator('main').evaluate((element) => {
     element.innerHTML = '<button>After action</button>'
   })
-  await recordTrace(page, task, 'after action', 1)
-})
-
-const recordTrace = vi.defineHelper(async (page: Page, task: Parameters<typeof recordArtifact>[0], name: string, startTime: number) => {
-  const snapshot = await page.evaluate(() => {
-    const { snapshot } = (globalThis as any).rrwebSnapshot
-    const serialized = snapshot(document)
-    if (!serialized) {
-      throw new Error('Failed to serialize document')
-    }
-    return {
-      serialized,
-      viewport: {
-        width: globalThis.innerWidth,
-        height: globalThis.innerHeight,
-      },
-      scroll: {
-        x: globalThis.scrollX,
-        y: globalThis.scrollY,
-      },
-      pseudoClassIds: {},
-    }
-  })
-
-  await recordArtifact(task, {
-    type: 'internal:browserTrace',
-    data: {
-      retry: 0,
-      repeats: 0,
-      recordCanvas: false,
-      entries: [{
-        name,
-        kind: 'mark',
-        startTime,
-        snapshot,
-      }],
-    },
-  })
+  await trace.snapshot('after action')
 })
