+++ test/ui/test/trace-custom.spec.ts
@@ -80,14 +80,14 @@ async function testCustomTrace(page: Page, baseURL: string) {
   await traceSteps.nth(0).click()
   const editor = page.getByTestId('editor')
   const activeLine = editor.locator('.CodeMirror-activeline')
-  await expect(activeLine).toHaveText(/await recordTrace\(page, task, 'before action', 0\)/)
+  await expect(activeLine).toHaveText(/await trace\.snapshot\('before action'\)/)
 
   const traceEditorMarkers = editor.getByTestId('trace-editor-marker')
   await expect(traceEditorMarkers).toHaveCount(2)
   await expect(traceEditorMarkers.nth(0)).toHaveAttribute('aria-current', 'step')
 
   await traceSteps.nth(1).click()
-  await expect(activeLine).toHaveText(/await recordTrace\(page, task, 'after action', 1\)/)
+  await expect(activeLine).toHaveText(/await trace\.snapshot\('after action'\)/)
   await expect(traceEditorMarkers.nth(1)).toHaveAttribute('aria-current', 'step')
   await expect(traceFrame.getByRole('button', { name: 'After action' })).toBeVisible()
 }
